% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_imputations.R
\name{how_many_imputations}
\alias{how_many_imputations}
\title{How Many Imputations?}
\usage{
how_many_imputations(model, cv = 0.05, alpha = 0.05)
}
\arguments{
\item{model}{Either a \code{mira} object (created by running a model on a
data set which was imputed using \code{mice}) or a \code{mipo} object
(creating by runing \code{pool()} on a \code{mira} object), or any object
which can be coverted to \code{mira} via \code{as.mira}.}

\item{cv}{Desired precision of standard errors. Default to .05. If the data
were re-imputed, the estimated standard errors would differ by no more than
this amount.}

\item{alpha}{Significance level for choice of "conservative" FMI.}
}
\value{
The number of required imputations to obtain the \code{cv} level of
  precision.
}
\description{
Implements two-stage "how_many_imputations" from von Hippel (2018)
}
\examples{
data(airquality)
# Add some missingness
airquality[4:10, 3] <- rep(NA, 7)
airquality[1:5, 4] <- NA
airquality <- airquality[-c(5, 6)]
impdata1 <- mice::mice(airquality, m = 5, maxit = 10, method = 'pmm', seed = 500)
modelFit1 <- with(impdata1, lm(Temp ~ Ozone + Solar.R + Wind))
how_many_imputations(modelFit1)
how_many_imputations(modelFit1, cv = .01)

# Using a non-`mice` library.
library(jomo)
library(mitools) # for the `imputationList` function
jomodata <- jomo::jomo1(airquality, nburn = 100, nbetween = 100, nimp = 5)
impdata2 <- mitools::imputationList(split(jomodata, jomodata$Imputation))
modelfit2 <- with(impdata2, lm(Temp ~ Ozone + Solar.R + Wind))
how_many_imputations(modelfit2)
}
\references{
von Hippel, Paul T. (2018)
\sQuote{How Many Imputations Do You Need? A Two-stage Calculation Using a Quadratic Rule.},
\emph{Sociological Methods & Research} p.0049124117747303.
}
