% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_imputations.R
\name{howManyImputations}
\alias{howManyImputations}
\title{Improper function name, use \code{how_many_imputations} instead}
\usage{
howManyImputations(model, cv = 0.05, alpha = 0.05)
}
\arguments{
\item{model}{See \code{how_many_imputations}}

\item{cv}{See \code{how_many_imputations}}

\item{alpha}{See \code{how_many_imputations}}
}
\value{
See \code{how_many_imputations}
}
\description{
Improper function name, use \code{how_many_imputations} instead
}
