% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_imputations.R
\name{how_many_imputations}
\alias{how_many_imputations}
\title{How Many Imputations?}
\usage{
how_many_imputations(model, cv = 0.05, alpha = 0.05)
}
\arguments{
\item{model}{Either a `mira` object (created by running a model on a data set which was imputed via "mice")
or a `mipo` object (creating by runing `pool()` on a `mira` object).}

\item{cv}{Desired precision of standard errors. Default to .05. (I.e., if the data were re-imputed, the
estimated standard errors would differ by no more than this amount.)}

\item{alpha}{Significance level for choice of "conservative" FMI.}
}
\value{
The number of required imputations to obtain the `cv` level of precision.
}
\description{
Implements two-stage "how_many_imputations" from von Hippel (2018)
}
\examples{
library(mice)
data <- airquality
data[4:10,3] <- rep(NA,7)
data[1:5,4] <- NA
data <- data[-c(5,6)]
tempData <- mice(data,m=5,maxit=10,meth='pmm',seed=500)
modelFit1 <- with(tempData,lm(Temp~ Ozone+Solar.R+Wind))
how_many_imputations(modelFit1)
how_many_imputations(pool(modelFit1))
how_many_imputations(modelFit1, cv = .01)
}
\references{
von Hippel, Paul T. (2018)
\sQuote{How Many Imputations Do You Need? A Two-stage Calculation Using a Quadratic Rule.},
\emph{Sociological Methods & Research} p.0049124117747303.
}
